requirements.txt
APScheduler==3.6.0
beautifulsoup4==4.9.0
Brotli==1.0.7
certifi==2019.3.9
chardet==3.0.4
click==7.1.2
csscompressor==0.9.5
Django==2.2
django-appconf==1.0.4
django-apscheduler==0.3.0
django-compressor==2.4
django-debug-toolbar==1.11
django-js-asset==1.2.1
django-jsonfield==1.2.0
django-mptt==0.9.1
django-redis==4.10.0
django-redis-cache==1.8.1
django-uuidfield==0.5.0
djangorestframework==3.9.1
esl==0.5.1
gTTS==2.2.1
gTTS-token==1.1.4
html5lib==1.1
idna==2.8
Jinja2==2.10.1
jsonfield==2.0.2
Louie==2.0
MarkupSafe==1.1.1
nose==1.3.7
numpy==1.16.1
pandas==0.24.1
ply==3.11
psycopg2==2.8.3
pycurl==7.43.0.2
pyodbc==4.0.30
python-dateutil==2.8.0
python-ESL==1.4.18
pytz==2018.9
pyzmq==2.2.0.1
rcssmin==1.0.6
redis==2.10.6
requests==2.21.0
rjsmin==1.1.0
six==1.12.0
socketIO-client-nexus==0.7.6
soupsieve==2.0
SQLAlchemy==1.3.2
sqlparse==0.3.0
threadpool==1.3.2
tzlocal==1.5.1
urllib3==1.24.1
webencodings==0.5.1
websocket-client==0.56.0
xlrd==1.2.0
XlsxWriter==1.3.7
xlutils==2.0.0
xlwt==1.3.0
